requirements.txt
fastapi
sqlalchemy
psycopg2-binary
pydantic-settings
uvicorn
pytest
pytest-mock
pytest-cov
flake8
black